+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * [alerts] Add support for the duration_before_trigger attribute [#407](https://github.com/Scalingo/cli/pull/407)
 * [domains] Handle Let's Encrypt certificate status [#410](https://github.com/Scalingo/cli/pull/410)
-* Fix login with SSH [#418](https://github.com/Scalingo/cli/pull/418)
+* Fix login with SSH [#419](https://github.com/Scalingo/cli/pull/419)
 
 ### 1.10.1